Fix #2299: Avoid trac blocking on pipebot.
Change to using pipebot as a submodule, rather than a local
implementation of writing to the file. This ensures that the right
things are passed when opening the file, and that it therefore
doesn't block.

Patch set 2: Log any exceptions which occur.
Patch set 3: Drop the submodule, copy the code instead. I've just
             spent a couple of hours playing with trac and eggs,
             and I can't make it work. While slightly nasty, we
             know that a single-file solution already works, so
             just use that instead.
Patch set 4: missed an import.

Change-Id: Ia1b4a239047674b21fb3cf101cf211bfcfcdbedd